fixing contest
