+++ frontend/src/context/CatalogReducer.js
@@ -109,12 +109,6 @@ const CatalogReducer = (state, action) => {
         contests: action.payload,
         loading: false,
       };
-    case 'FETCH_CONTEST':
-      return {
-        ...state,
-        contest: action.payload,
-        loading: false,
-      };
     case 'CREATE_CONTEST':
       return {
         ...state,
