+++ frontend/src/context/ContestContext.js
@@ -6,10 +6,6 @@ const ContestContext = createContext({
     category: {},
     categoryWods: [],
     categoryWod: {},
-    fetchContest:  async () => {},
-    createContest:  async () => {},
-    updateContest:  async () => {},
-    deleteContest:  async () => {},
     addCategory:  async () => {},
     deleteCategory:  async () => {},
     addCategoryWod:  async () => {},
