+++ frontend/src/context/ContestReducer.js
@@ -1,7 +1,6 @@
 const contestReducer = (state, action) => {
     switch (action.type) {
       case 'FETCH_CONTEST':
-        console.log("Entry aqui")
         return {
           ...state,
           categories: action.payload?.contestCategory?.map(category => {
@@ -17,27 +16,6 @@ const contestReducer = (state, action) => {
               }
             })
           },
-          loading: false,
-        };
-      case 'CREATE_CONTEST':
-        return {
-          ...state,
-          contest: action.payload,
-          contests: [...state.contests, action.payload],
-          loading: false,
-        };
-      case 'UPDATE_CONTEST':
-        return {
-          ...state,
-          contest: action.payload,
-          contests: state.contests.map((contest) =>
-            contest.id === action.payload.id ? action.payload : contest,
-          ),
-        };
-      case 'DELETE_CONTEST':
-        return {
-          ...state,
-          contest: action.payload,
           loading: false,
         };
       case 'ADD_CATEGORY':
