+++ frontend/src/hooks/useContest.js
@@ -18,14 +18,7 @@ const useContest = (dispatch) => {
   const setLoading = (loading) => {
     loadingDispatch({ type: 'SET_LOADING', payload: loading });
   };
-  const fetchContests = useMutation({
-    mutationFn: getContests,
-    onMutate: () => setLoading(true),
-    onSuccess: (data) => {
-      dispatch({ type: "FETCH_CONTESTS", payload: data });
-    },
-    onSettled: () => setLoading(false),
-  });
+
 
   const fetchContest = useMutation({
     mutationFn: getContest,
@@ -37,64 +30,6 @@ const useContest = (dispatch) => {
     onSettled: () => setLoading(false),
   });
 
-  const useCreateContest = useMutation({
-    mutationFn: createContest,
-    onMutate: () => {
-      setLoading(true);
-    },
-    onSuccess: (data) => {
-      dispatch({ type: 'CREATE_CONTEST', payload: data });
-      Notifies('success', 'Competencia creada correctamente');
-    },
-    onError: (error) => {
-      console.log('error on createContest', error);
-      Notifies('error', error?.response?.data?.message);
-      setLoading(false);
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries('contests');
-      setLoading(false);
-    },
-  });
-
-  const useUpdateContest = useMutation({
-    mutationFn: updateContest,
-    onMutate: () => {
-      setLoading(true);
-    },
-    onSuccess: (data) => {
-      dispatch({ type: 'UPDATE_CONTEST', payload: data });
-      Notifies('success', 'Competencia actualizada correctamente');
-    },
-    onError: (error) => {
-      console.log('error on updateContest', error);
-      setLoading(false);
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries('contests');
-      setLoading(false);
-    },
-  });
-
-
-  const useDeleteContest = useMutation({
-    mutationFn: deleteContest,
-    onMutate: () => {
-      setLoading(true);
-    },
-    onSuccess: (data) => {
-      dispatch({ type: 'DELETE_CONTEST', payload: data });
-      Notifies('success', 'Competencia eliminada correctamente');
-    },
-    onError: (error) => {
-      console.log('error on deleteContest', error);
-      setLoading(false);
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries('contests');
-      setLoading(false);
-    },
-  });
   // Contest Categories
   const useAddCategory = useMutation({
     mutationFn: addCategory,
@@ -135,16 +70,6 @@ const useContest = (dispatch) => {
   });
   return {
     fetchContest: fetchContest.mutate,
-    fetchContests: fetchContests.mutate,
-    createContest: (values) => {
-      return useCreateContest.mutateAsync(values);
-    },
-    updateContest: (values) => {
-      return useUpdateContest.mutateAsync(values);
-    },
-    deleteContest: (values) => {
-      return useDeleteContest.mutateAsync(values);
-    },
     addCategory: (values) => {
       return useAddCategory.mutateAsync(values)
     },
